requirements.txt
fastapi==0.85.0
uvicorn[standard]
psycopg2-binary==2.9.4
SQLAlchemy==1.4.41
pytest==7.1.3
pytest-schema==0.1.1
requests==2.28.1
flake8==5.0.4
Faker==15.0.0
python-multipart==0.0.5
Pillow==9.2.0
